Smooth the motion along X or along Y before stop
If the distance in XY plane is not enough to stop the motion then we go backwards through the command buffer and slow down the motion in the previous commands. This recursive procedure goes on until overall distance in XY plane is enough to stop motion keeping the default acceleration.
So far this function is tested only separately along X or Y axes. More tests for the general XY motion are required.